refactor: [unit tests] replace test-data to separate files
