+++ src/components/cross-check-session-description/test-data.js
@@ -1,0 +1,19 @@
+const initialValues = {
+  id: "xcheck-1",
+  title: "rss2020Q3react-songbird",
+  author: "viktorsipach",
+  state: "draft",
+  taskTitle: "Task 1",
+  crossCheckSessionPeriod: [
+    "2020-07-08",
+    "2020-07-14"
+  ],
+  coefficient: 0.2,
+  discardMinScore: true,
+  discardMaxScore: true,
+  minReviewsAmount: 1,
+  desiredReviewsAmount: 2,
+  attendees: []
+}
+
+export default initialValues;
