+++ src/components/cross-check-sessions-table/cross-check-sessions-table.test.js
@@ -2,43 +2,7 @@ import React from 'react';
 import {render} from "@testing-library/react";
 import CrossCheckSessionsTable from './cross-check-sessions-table';
 import mapData from './map-data';
-
-const tableData = [
-  {
-    "id": "xcheck-1",
-    "title": "rss2020Q3react-songbird",
-    "author": "viktorsipach",
-    "state": "draft",
-    "taskTitle": "Task 1",
-    "crossCheckSessionPeriod": [
-      "2020-07-08",
-      "2020-07-14"
-    ],
-    "coefficient": 0.2,
-    "discardMinScore": true,
-    "discardMaxScore": true,
-    "minReviewsAmount": 1,
-    "desiredReviewsAmount": 2,
-    "attendees": []
-  },
-  {
-    "id": "xcheck-2",
-    "title": "rss2020Q3react-xcheck",
-    "author": "KatiaR",
-    "state": "draft",
-    "taskTitle": "Task 2",
-    "crossCheckSessionPeriod": [
-      "2020-09-15",
-      "2020-09-17"
-    ],
-    "coefficient": 0.7,
-    "discardMinScore": true,
-    "discardMaxScore": true,
-    "minReviewsAmount": 1,
-    "desiredReviewsAmount": 2,
-    "attendees": []
-  }
-]
+import tableData from './test-data';
 
 describe('map data', () => {
   test('map data', () => {
